+++ bipred/__init__.py
@@ -40,7 +40,7 @@ from ldpred3's ``[fast]`` extra.
 
 import importlib
 
-__version__ = "0.1.0.dev0"
+__version__ = "0.1.0.dev1"
 
 # public name -> submodule it lives in
 _EXPORTS = {